lab2 fixed

+++ src/itmo/java/basics/lab2/Calculator.java
@@ -25,16 +25,16 @@ public class Calculator {
         return a*b;
     }
 
-    public int div (int a, int b) {
-        return a/b;
+    public double div (int a, int b) {
+        return a/(double)b;
     }
 
     public double div (double a, double b) {
         return a/b;
     }
 
-    public long div (long a, long b) {
-        return a/b;
+    public float div (long a, long b) {
+        return a/(float)b;
     }
 
     public int sub (int a, int b) {
